% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{calculate_baseline}
\alias{calculate_baseline}
\title{Calculate baseline for normalization}
\usage{
calculate_baseline(wd, beads.gates, beads.type)
}
\arguments{
\item{wd}{Working directory (character)}

\item{beads.gates}{Gates to identify the beads. This is a data structure with the following format
\code{list(file_name = list(channel_name = list(x = [xMin, xMax], y = [yMin, yMax]), ...), ...)}.
Note that only files in \code{names(beads.gates)} will be processed. Also note that the data structure
may contain gates for extra channels, but which channels will be used depends on the \code{beads.type} parameter}

\item{beads.type}{Type of beads. Must be on of \code{"Fluidigm"}, \code{"Beta"}}
}
\value{
Returns a vector with the medians of the beads events across all processed files
}
\description{
Extract beads events from all the FCS files in the working directory
and calculates the baseline for normalization
}

