% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{calculate_bcs_separation}
\alias{calculate_bcs_separation}
\title{This assumes that all barcodes are identified by the same number of channels}
\usage{
calculate_bcs_separation(m, bc.channels, expected.positive, cutoff)
}
\arguments{
\item{m}{Transformed data matrix}

\item{expected.positive}{A single number. The expected number of positive barcode channels for each event}
}
\description{
This assumes that all barcodes are identified by the same number of channels
}
