% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{correct_data_channels}
\alias{correct_data_channels}
\title{Normalize data}
\usage{
correct_data_channels(m, beads.data, baseline, beads.col.names,
  time.col.name = "Time")
}
\arguments{
\item{m}{The data matrix}

\item{beads.data}{The beads.data for this data matrix. This would typically be a subset of the rows
representing the bead events}

\item{baseline}{A vector of values containing the bead channels values to be used as baseline
for normalization}

\item{beads.col.names}{A vector with the names of the beads channels}

\item{time.col.name}{The name of the time column}
}
\value{
Returns a list with the following components
\itemize{
 \item \code{m.normed} The normalized data matrix
 \item \code{smoothed.beads} A matrix with the smoothed beads data
 \item \code{beads.slopes} A matrix with the calculated beadlsopes with \code{nrow(beads.slopes) = nrow(beads.data)} and
 two columns: \code{time} and \code{slope}
}
}
\description{
This function performs the actual normalization
}

