% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cytofNormalizeR}
\alias{cytofNormalizeR}
\alias{cytofNormalizeR-package}
\title{cytofNormalizeR: A package for normalization and debarcoding of CyTOF data.}
\description{
To start the GUI refer to the documentation for the \code{cytofNormalizeR.run}
by typing \code{help('cytofNormalizeR.run')} in your R session
}

