% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{debarcode_data}
\alias{debarcode_data}
\title{Debarcodes data by doing normalization after preliminary barcode assignemnts}
\usage{
debarcode_data(m, bc.key)
}
\description{
Debarcodes data by doing normalization after preliminary barcode assignemnts
}

