% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{debarcode_data_matrix}
\alias{debarcode_data_matrix}
\title{Debarcodes an individual data matrix}
\usage{
debarcode_data_matrix(m, bc.channels, bc.key)
}
\description{
Debarcodes an individual data matrix
}

