% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{debarcoder_GUI}
\alias{debarcoder_GUI}
\title{Starts the debarcoder GUI}
\usage{
debarcoder_GUI(...)
}
\arguments{
\item{...}{Additional arguments to be passed to \code{shiny::runApp}}
}
\description{
To stop the software simply hit the "ESC" key in your R session.
}
