% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_io.R
\name{flowCore_read_FCS}
\alias{flowCore_read_FCS}
\title{flowCore_read_FCS}
\usage{
flowCore_read_FCS(
  filename,
  emptyValue = FALSE,
  transformation = FALSE,
  truncate_max_range = FALSE,
  ...
)
}
\arguments{
\item{filename}{name of the FCS file}

\item{emptyValue}{set to FALSE, see original documentation}

\item{transformation, }{set to FALSE, ie no transformation}

\item{truncate_max_range}{set to FALSE, ie no truncation}

\item{...}{parameters passed to read.FCS}
}
\value{
a flowFrame
}
\description{
This function calls flowCore read.FCS function, but with modified default
arguments
}
