% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{get_assignments_at_threshold}
\alias{get_assignments_at_threshold}
\title{Assign events to a specific sample}
\usage{
get_assignments_at_threshold(bc.results, sep.threshold,
  mahal.threshold = NULL, mahal.dist = NULL)
}
\arguments{
\item{bc.results}{The debarcoding results returned from \code{debarcode_data}}

\item{sep.threshold}{The threshold value for the separation between the positive and negative barcode
channels. Only events whose separation is greater than this threshold are assigned to the sample}

\item{mahal.threshold}{If not \code{NULL} The threshold value for the Mahalanobis distance between each
event and the centroid of the sample the event has been assigned to. Only events with distance lower
than this threshold are assigned to the sample.}

\item{mahal.dist}{A vector of Mahalnobis distances, as returned by \code{get_mahalanobis_distance}. Only
used if \code{mahal.threshold} is not \code{NULL}}
}
\value{
Returns a character vector containing the sample labels, given the current thresholds.
}
\description{
Assign events to a specific sample
}

