% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{get_barcode_channels_names}
\alias{get_barcode_channels_names}
\title{Get the names of the barcode channels}
\usage{
get_barcode_channels_names(m, bc.key)
}
\arguments{
\item{m}{The data matrix, as returned by \code{flowCore::exprs}}

\item{bc.key}{The barcode key, as returned by \code{read_barcode_key}}
}
\value{
Returns a vector containing the names of the barcode channels. The vector can be used as index
 to select the appropriate columns of \code{m}
}
\description{
Get the names of the barcode channels
}
