% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{get_mahalanobis_distance}
\alias{get_mahalanobis_distance}
\title{Calculates the squared Mahalanobis distance from the centroid of the barcoded sample}
\usage{
get_mahalanobis_distance(m, bc.res, sep.threshold)
}
\arguments{
\item{m}{The data matrix}

\item{bc.res}{The debarcoding results}

\item{sep.threshold}{The minimum separation in the intensity of the barcoding channels. Barcode assignments
are performed based on this threshold. Then the function calculates the Mahalanobis distance between each event
and the centroid of the debarcoded sample, as assigned using this threshold.}
}
\value{
Returns a vector of squared Mahalanobis distances, as calculated by \code{stats::mahalanobis}. The distance
 values are capped at 30
}
\description{
Calculates the squared Mahalanobis distance from the centroid of the barcoded sample
}

