% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{get_mahalanobis_distance_from_beads}
\alias{get_mahalanobis_distance_from_beads}
\title{Calculates the Mahalanobis distance of each event from the centroid of the beads population}
\usage{
get_mahalanobis_distance_from_beads(m, beads.cols.names)
}
\arguments{
\item{m}{The untransformed matrix of data. Must contain a column called \code{beadEvent} which indicates
when the corresponding event is a bead (1) or not (0)}

\item{beads.cols.names}{The names of the bead columns}
}
\value{
A vector of length \code{nrow(m)} containing the \code{sqrt} of the Mahalanobis distance of each of
\code{m} from the centroid of the beads population
}
\description{
Calculates the Mahalanobis distance of each event from the centroid of the beads population
}

