% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{get_sample_idx}
\alias{get_sample_idx}
\title{Returns the indices of the events corresponding to a given sample label}
\usage{
get_sample_idx(
  label,
  bc.results,
  sep.threshold,
  mahal.threshold = NULL,
  mahal.dist = NULL
)
}
\arguments{
\item{label}{Character string. The desired sample label}

\item{bc.results}{The debarcoding results returned from \code{debarcode_data}}

\item{sep.threshold}{The threshold value for the separation between the positive and negative barcode
channels. Only events whose separation is greater than this threshold are assigned to the sample}

\item{mahal.threshold}{If not \code{NULL} The threshold value for the Mahalanobis distance between each
event and the centroid of the sample the event has been assigned to. Only events with distance lower
than this threshold are assigned to the sample.}

\item{mahal.dist}{A vector of Mahalnobis distances, as returned by \code{get_mahalanobis_distance}. Only
used if \code{mahal.threshold} is not \code{NULL}}
}
\value{
Returns a boolean vector with \code{TRUE} corresponding to the desired events
}
\description{
Returns the indices of the events corresponding to a given sample label
}
