% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{identify_beads}
\alias{identify_beads}
\title{Beads identification}
\usage{
identify_beads(m, gates, beads.cols.names, dna.col)
}
\arguments{
\item{m}{The matrix, asinh transformed}

\item{gates}{Gates to be applied (only the gates for this specific matrix)}

\item{beads.cols.names}{A character vector containing the names of the beads columns}

\item{dna.col}{The index (either character, numeric or boolean) of the DNA channel column}
}
\value{
Returns a boolean vector of length \code{nrow(m)} with the indexes of the bead events
}
\description{
Applies the beads gates to the data matrix and idenfies the beads
}

