% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{normalize_folder}
\alias{normalize_folder}
\title{Normalize a folder of FCS files}
\usage{
normalize_folder(wd, output.dir.name, beads.gates, beads.type,
  baseline = NULL)
}
\arguments{
\item{wd}{Working directory (character)}

\item{output.dir.name}{Name of the output directory (character). The output directory
will be created as a subfolder of the Working directory}

\item{beads.gates}{Gates to identify the beads. This is a data structure with the following format
\code{list(file_name = list(channel_name = list(x = [xMin, xMax], y = [yMin, yMax]), ...), ...)}.
Note that only files in \code{names(beads.gates)} will be processed. Also note that the data structure
may contain gates for extra channels, but which channels will be used depends on the \code{beads.type} parameter}

\item{beads.type}{Type of beads. Must be on of \code{"Fluidigm"}, \code{"Beta"}}

\item{baseline}{If \code{NULL} the median beads intensities of the current files will be used as baseline
for normalization. Alternatively this can be a character string with the path of a directory containing
FCS files of beads events, whose median intensities will be used as baseline.}
}
\description{
Main wrapper for the normalization pipeline. Use beads gates to identify beads, calculate a baseline
for normalization and then normalize the files, writing new output FCS
}
