% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{paneleditor_GUI}
\alias{paneleditor_GUI}
\title{Starts the panel editor GUI}
\usage{
paneleditor_GUI(launch.browser = TRUE, ...)
}
\arguments{
\item{launch.browser}{Whether to open the GUI in a separate browser window
(recommended)}

\item{...}{Additional arguments to be passed to \code{shiny::runApp}}
}
\description{
Upon starting, a file selection window will appear from your R session. You should use
this window to navigate to the directory containing the data you want to analyze,
and select any file in that directory. The directory itself will then become
the working directory for the software.
To stop the software simply hit the "ESC" key in your R session.
}
