% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder_plotting.R
\name{plot_all_barcode_biaxials}
\alias{plot_all_barcode_biaxials}
\title{Scatteplot matrix of all the barcode channels}
\usage{
plot_all_barcode_biaxials(m, bc.channels)
}
\arguments{
\item{m}{The matrix of data to plot (\code{asinh} transformed)}

\item{bc.channels}{The names of the barcode channels}
}
\value{
Returns a ggplot2 object
}
\description{
Scatteplot matrix of all the barcode channels
}

