% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder_plotting.R
\name{plot_barcode_channels_intensities}
\alias{plot_barcode_channels_intensities}
\title{Plotting barcode channels intensities for individual events}
\usage{
plot_barcode_channels_intensities(m, bc.channels, m.normed = NULL)
}
\arguments{
\item{m}{The data matrix to plot. The value should be \code{asinh} transformed}

\item{bc.channels}{The barcode channels names}

\item{m.normed}{A matrix of rescaled intensity values. If provided, they are plotted
next to the values in \code{m}}
}
\value{
Returns a ggplot object
}
\description{
Plotting barcode channels intensities for individual events
}

