% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder_plotting.R
\name{plot_barcode_separation}
\alias{plot_barcode_separation}
\title{Plotting debarcoded sample yields as a function of the separation threshold}
\usage{
plot_barcode_separation(bc.res, sep.threshold = NULL)
}
\arguments{
\item{bc.res}{The debarcoding results to plot}

\item{sep.threshold}{If provided, adds a red line corresponding to the separation
threshold in use}
}
\value{
Returns a ggplot object
}
\description{
Plotting debarcoded sample yields as a function of the separation threshold
}

