% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer_plotting.R
\name{plot_distance_from_beads}
\alias{plot_distance_from_beads}
\title{Biaxial plot color-coded according to the Mahalanobis distance from the beads population}
\usage{
plot_distance_from_beads(m, x.var, y.var)
}
\arguments{
\item{m}{The untrasformed matrix of data. Must contain a column named \code{beadDist} containing the
Mahalanobis distance}

\item{x.var}{The variable on the x axis}

\item{y.var}{The variable on the y axis}
}
\description{
Biaxial plot color-coded according to the Mahalanobis distance from the beads population
}
