% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder_plotting.R
\name{plot_separation_histogram}
\alias{plot_separation_histogram}
\title{Plotting the distribution of separation values}
\usage{
plot_separation_histogram(bc.results)
}
\arguments{
\item{bc.results}{The debarcoding results}
}
\value{
Returns a ggplot object
}
\description{
Plotting the distribution of separation values
}

