% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{premessa}
\alias{premessa}
\title{premessa: A package for normalization and debarcoding of CyTOF data.}
\description{
For the normalizer GUI refer to the documentation for the \code{normalizer_GUI} function
by typing \code{help('normalizer_GUI')} in your R session. For the debarcoder GUI refer
to the documentation for the \code{debarcoder_GUI} function by typing \code{help('debarcoder_GUI')} in your R session
}
