% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{quickbarcoder_GUI}
\alias{quickbarcoder_GUI}
\title{Starts the quick barcoder GUI}
\usage{
quickbarcoder_GUI(launch.browser = TRUE, ...)
}
\description{
Upon starting, a file selection window will appear from your R session. You should use
this window to navigate to the directory containing the data you want to analyze,
and select any file in that directory. The directory itself will then become
the working directory for the software.
To stop the software simply hit the "ESC" key in your R session.
}
