% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{read_barcode_key}
\alias{read_barcode_key}
\title{Loading a barcode key}
\usage{
read_barcode_key(fname)
}
\arguments{
\item{fname}{The path of the file containing the barcode key. Please refer to the online documentation
for a description of the format}
}
\value{
Returns a data frame containing the barcode key
}
\description{
Loading a barcode key
}
