% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneleditor_helpers.R
\name{read_parameters}
\alias{read_parameters}
\title{Read parameters from a set of FCS files}
\usage{
read_parameters(files.list)
}
\arguments{
\item{files.list}{A character vector containing the paths to the files}
}
\value{
Returns a data.frame with \code{row.names} corresponding to the union of all
  parameter names (i.e. $PnN) in all the files. Each column correspond to a file
  and the \code{[row, column]} values indicating the description (i.e. $PnS) value
  for the \code{row} parameter in the \code{column} file. If the description field
  is empty, the value is \code{""}. If a parameter is missing in a file the value
  is \code{NA}
}
\description{
This function reads information about the parameters in a set of FCS files
and returns the aggregate information in a table
}

