% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{remove_beads_from_file}
\alias{remove_beads_from_file}
\title{Remove beads from an FCS file}
\usage{
remove_beads_from_file(input.fname, dist.threshold, out.dir)
}
\arguments{
\item{input.fname}{The path to the FCS file from which beads must be removed}

\item{dist.threshold}{The Mahalanobis distance threhsold for beads removal. Events with
\code{beadDist <= dist.threshold} will be removed}

\item{out.dir}{The path to the output directory}
}
\description{
This function removes beads from an FCS file according to a specified Mahalanobis distance cutoff
and writes the results in an output folder
}
