% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneleditor_helpers.R
\name{rename_fcs_parameters_desc}
\alias{rename_fcs_parameters_desc}
\title{Rename FCS parameter descriptions}
\usage{
rename_fcs_parameters_desc(fcs, names.map)
}
\arguments{
\item{fcs}{FCS data as returned by \code{read_fcs}}

\item{names.map}{A named character vector mapping parameter names (i.e. $PnN) to
the desired parameter descriptions ($PnS). The \code{names} of this vector
should be the parameter names, and the values the parameter descriptions. All
the parameter names of the \code{fcs} data must be present in \code{names(names.map)}}
}
\value{
Returns the FCS data with the updated descriptions
}
\description{
This function renames FCS parameter descriptions (also called "short names", i.e.
  the $PnS keyword)
}
