% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneleditor_helpers.R
\name{rename_fcs_parameters_name}
\alias{rename_fcs_parameters_name}
\title{Rename FCS parameter names}
\usage{
rename_fcs_parameters_name(fcs, names.map)
}
\arguments{
\item{fcs}{FCS data as returned by \code{read_fcs}}

\item{names.map}{A named character vector mapping old names to new names.
The \code{names} of this vector should be the old names, and the values
the corresponding new names. New names need to be unique}
}
\value{
Returns the FCS data with the updated names
}
\description{
This function renames FCS parameter names (i.e. the $PnN keyword)
}

