% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneleditor_helpers.R
\name{rename_parameters_in_files}
\alias{rename_parameters_in_files}
\title{Renames FCS parameters in a set of files}
\usage{
rename_parameters_in_files(working.dir, out.dir, tab)
}
\arguments{
\item{working.dir}{The directory in which the files are located}

\item{out.dir}{The output directory. This will be created as a subfolder of
\code{working.dir}}

\item{tab}{The table specifying how the renaming should be performed. This
 is a structure similar as the \code{data.frame} returned from \code{read_parameters}. \code{row.names}
 correspond to the current parameter names in the files, and columns correspond to files (which
 should be located in the \code{working.dir}). The value at \code{[row, column]} should contain the
 value you want to set for the corresponding parameter description (i.e. $PnS). In addition the
 \code{data.frame} must contain two extra columns with the following names
 \itemize{
    \item{\code{Remove}}{ logical indicating whether the parameter should be removed from the files}
    \item{\code{Parameter}}{ character indicating the desired new name of the parameter (i.e. the new
    $PnN)}
}}
}
\description{
This function renames FCS parameters in a set of files, as specified by a template
}
