% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{write_FCS}
\alias{write_FCS}
\title{Write FCS files, with proper processing of parameters and keywords}
\usage{
write_FCS(exprs.m, out.name, source.flowFrame = NULL)
}
\arguments{
\item{exprs.m}{The data matrix. Parameter names will be taken from \code{colnames(exprs.m)},
but see also the description of the \code{source.flowFrame} parameter}

\item{out.name}{The name of the output file}

\item{source.flowFrame}{If a flowFrame object is supplied, the function will copy matching names and descriptions
keywords from it (e.g. $P1S in the \code{source.flowFrame} is copied to $P1S of the new flowFrame etc.). Extra columns present
in exprs.m are preserved (i.e. if source.flowFrame doesn't contain $P1S the original version is preserved)}
}
\description{
Write FCS files, with proper processing of parameters and keywords
}

