% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_io.R
\name{write_flowFrame}
\alias{write_flowFrame}
\title{Write a flowFrame as FCS file}
\usage{
write_flowFrame(flowFrame, path)
}
\arguments{
\item{flowFrame}{the \code{flowFrame} to write}

\item{path}{destination path}
}
\description{
This function writes a flowFrame as an FCS file, taking care of updating the \code{$FILENAME} keyword
}
