% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{debarcode_fcs}
\alias{debarcode_fcs}
\title{Main wrapper for the debarcoder pipeline. Takes all the parameters as input
does everything and writes back the FCS files}
\usage{
debarcode_fcs(fcs, bc.key, output.dir, output.basename, sep.threshold)
}
\description{
Main wrapper for the debarcoder pipeline. Takes all the parameters as input
does everything and writes back the FCS files
}

