% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcode_cytof.R
\name{normalize_by_pop}
\alias{normalize_by_pop}
\title{Rows for which bc.labels is NA are left unnormalized}
\usage{
normalize_by_pop(m, bc.channels, bc.labels)
}
\description{
Rows for which bc.labels is NA are left unnormalized
}

