% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cytof.R
\name{remove_beads_from_fcs}
\alias{remove_beads_from_fcs}
\title{Remove beads from a flowFrame object}
\usage{
remove_beads_from_fcs(fcs, dist.threshold)
}
\arguments{
\item{fcs}{The \code{flowFrame} object from which beads must be removed. Must contain a column
called \code{beadDist} representing the Mahalanobis distance of each event from the centroid
of the beads population}

\item{dist.threshold}{The Mahalanobis distance threhsold for beads removal. Events with
\code{beadDist <= dist.threshold} will be removed}
}
\value{
Returns a list of two \code{flowFrame} objects
\itemize{
 \item{\code{data.fcs}}{ The data without the beads events}
 \item{\code{beads.fcs}}{ The beads events}
}
}
\description{
Remove beads from a flowFrame object
}

